- 🔥 Removed unused code across the project. Farewell, altGreetings.js, dropdown UI, and their posse – you will code no more. - 📦 Deleted styles and services, because who needs redundant baggage in life? - 🎨 Streamlined project: no more wasted bytes, no more JavaScript ghost stories. - 🧹 Cleanup so thorough, Marie Kondo would be proud.
