+++ src/css/charactersList.css
