--- src/css/forms.css
@@ -1,25 +1,0 @@
-/* ===== FORM ELEMENTS ===== */
-#acm_personality_textarea {
-    width: 100%;
-}
-
-#acm_scenario_pole {
-    width: 100%;
-    margin-left: 0;
-}
-
-#acm_talkativeness_div input[type="range"] {
-    width: 100%;
-}
-
-#acm_mes_example_div {
-    height: 100%;
-    display: flex;
-    flex-grow: 1;
-}
-
-#acm_mes_example_textarea {
-    width: 100%;
-    height: 100%;
-    margin-left: 0;
-}
