+++ src/css/modal.css
@@ -117,3 +117,61 @@
     display: none;
     z-index: 9100;
 }
+
+/* ===== DROPDOWN COMPONENTS ===== */
+.dropdown-ui-menu {
+    display: none;
+    position: absolute;
+    z-index: 9000;
+}
+
+.dropdown-ui-item {
+    padding: 8px 12px;
+    cursor: pointer;
+}
+
+.dropdown-ui-item:hover {
+    background-color: var(--white50a);
+    color: var(--black70a);
+}
+
+.dropdown-container {
+    width: 100%;
+    margin: 0 auto;
+    border-radius: 4px;
+    overflow: hidden;
+}
+
+.dropdown-title {
+    padding: 15px;
+    cursor: pointer;
+    font-weight: bold;
+    text-align: left;
+    position: relative;
+}
+
+.dropdown-content {
+    max-height: 0;
+    overflow: hidden;
+    transition: max-height 0.5s ease;
+    padding: 0 15px;
+}
+
+.dropdown-container.open .dropdown-content {
+    max-height: fit-content;
+    padding-top: 10px;
+    padding-bottom: 10px;
+    height: 100%;
+    overflow-y: auto;
+}
+
+.dropdown-submenu-item {
+    padding: 8px 12px;
+    cursor: pointer;
+}
+
+.dropdown-separator {
+    border: none;
+    border-top: 1px solid #ccc;
+    margin: 8px 0;
+}
