+++ src/services/characters-service.js
@@ -1,6 +1,5 @@
 import {
     getPastCharacterChats,
-    reloadCurrentChat,
     setCharacterId,
     system_message_types
 } from '../../../../../../script.js';
@@ -18,12 +17,13 @@ import {
     getThumbnailUrl,
     saveSettingsDebounced,
     extensionSettings,
-    getCharacters
+    getCharacters,
+    getTokenCountAsync,
+    substituteParams
 } from "../constants/context.js";
 import { debounce, delay } from '../utils.js';
 import { renameTagKey } from './tags-service.js';
-import { selectAndDisplay } from "../components/characters.js";
-import {setSelectedChar} from "../constants/settings.js";
+import {selectedChar, setSelectedChar} from "../constants/settings.js";
 
 /**
  * Checks the availability of the AvatarEdit API by making a POST request to the probe endpoint.
@@ -325,4 +325,48 @@ export async function exportChar(format, avatar) {
         document.body.removeChild(a);
     }
     $('#acm_export_format_popup').hide();
+}
+
+/**
+ * Collects the values of all textareas with the class 'altGreeting_zone'
+ * and returns them as an array of strings.
+ *
+ * @return {string[]} An array containing the values of the textareas with the class 'altGreeting_zone'.
+ */
+function generateGreetingArray() {
+    const textareas = document.querySelectorAll('.altGreeting_zone');
+    const greetingArray = [];
+
+    textareas.forEach(textarea => {
+        greetingArray.push(textarea.value);
+    });
+    return greetingArray;
+}
+
+/**
+ * Saves alternate greetings for the selected character and updates the relevant UI elements.
+ *
+ * @param {Event|null} event - The event object triggered by a user action, used to update token count.
+ *                             Pass null if no event is available.
+ * @return {void} This function does not return a value.
+ */
+export async function saveAltGreetings(event = null){
+    const greetings = generateGreetingArray();
+    const update = {
+        avatar: selectedChar,
+        data: {
+            alternate_greetings: greetings,
+        },
+    };
+    editCharDebounced(update);
+
+    // Update token count if necessary
+    if (event) {
+        const textarea = event.target;
+        const tokensSpan = textarea.closest('.inline-drawer-content').previousElementSibling.querySelector('.tokens_count');
+        tokensSpan.textContent = `Tokens: ${await getTokenCountAsync(substituteParams(textarea.value))}`;
+    }
+
+    // Edit the Alt Greetings number on the main drawer
+    $('#altGreetings_number').html(`Numbers: ${greetings.length}`);
 }
