+++ src/utils.js
@@ -46,9 +46,9 @@ export async function resetScrollHeight(element) {
 }
 
 /**
- * Delays the current async function by the given amount of milliseconds.
+ * Delays the current async function by the given number of milliseconds.
  * @param {number} ms Milliseconds to wait
- * @returns {Promise<void>} Promise that resolves after the given amount of milliseconds
+ * @returns {Promise<void>} Promise that resolves after the given number of milliseconds
  */
 export function delay(ms) {
     return new Promise(resolve => setTimeout(resolve, ms));
